Add flaky decorator to both estimateGas tests
